Os dependent third party lib install links. Extracted lib renaming.

+++ third/bin/main.py
@@ -4,15 +4,17 @@ from data import Data
 import link
 
 def main() -> None:
-    Data.init()
+    if not Data.init():
+        print("Data initialization failed")
+        return
 
     if (os.path.exists(f"{Data.projDir}/{Data.libDir}")):
         shutil.rmtree(f"{Data.projDir}/{Data.libDir}")
 
     os.makedirs(f"{Data.projDir}/{Data.libDir}")
 
-    for l in Data.libLinks:
-        link.get(l, f"{Data.projDir}/{Data.libDir}", tar=True)
+    for lib in Data.libs:
+        link.get(lib)
 
     return
